-ModelosSuperiores tiene un map en lugar de un vector. Es casi una copia de Modelo1. - Corregidos erroes de clase PPMC. Adaptada a las nuevas estucturas.

+++ Kaggle/fiuba.datos.ppmc/Modelo0.cpp
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-Modelo0::Modelo0(unsigned primo){
+Modelo0::Modelo0(unsigned long primo){
 	this->unHash = new Jenkins(primo);
 	this->unMapa = new MapaFrecuencia();
 	//this->palabras = new vector<Palabra*>;
